requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
dj-database-url==0.4.2
Django==1.8.15
django-background-tasks==1.1.13
django-compat==1.0.15
gunicorn==19.6.0
idna==2.6
igdb-api-python==1.0
psycopg2==2.6.2
requests==2.18.4
simplejson==3.13.2
six==1.11.0
urllib3==1.22
whitenoise==3.2.3
